backend/requirements.txt
asgiref==3.2.10
django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
gunicorn==20.0.4
Pillow==9.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.19.2
pytest-pythonpath==0.7.3
reportlab==3.6.9
sqlparse==0.3.1